requirements.txt
scrapy
beautifulsoup4
requests
twilio
PyYAML
APScheduler
python-telegram-bot
requests
selenium